Added More Exception

+++ src/main/java/com/capstone/Service/CustomerService.java
@@ -1,5 +1,6 @@
 package com.capstone.Service;
 
+import com.capstone.Exception.ResourseNotFound;
 import com.capstone.Model.Customer;
 import com.capstone.Repository.CustomerRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,19 +19,27 @@ public class CustomerService {
         return repo.save(cst);
     }
 
-    public Optional<Customer> findCustomer(Integer cstId) throws Exception
+    public Optional<Customer> findCustomer(Integer cstId) throws ResourseNotFound
     {
         if(repo.findById(cstId).isEmpty())
         {
             System.out.println("Reaching");
-            throw new Exception();
+            throw new ResourseNotFound("String is Empty");
 
         }
         return repo.findById(cstId);
     }
 
-    public void deleteCustomer(Integer cstId)
-    {System.out.println("Inside delete repo");
+    public void deleteCustomer(Integer cstId) throws ResourseNotFound
+    {
+        if(repo.findById(cstId).isEmpty())
+        {
+            //System.out.println("");
+            throw new ResourseNotFound("No Item is there to Deleted");
+
+        }
+
+        System.out.println("Inside delete repo");
         repo.deleteById(cstId);
     }
